# schemas/api/openapi.yaml
openapi: 3.1.0
info:
  title: ShakerProxy Control API
  version: 0.1.0-dev
  description: |
    Setup-safe API. Only operations implemented by this build are listed.

    Discover resources with GET /api/v1 and fetch this document from GET /api/v1/openapi.yaml.
    Every error, including unknown routes (404 not_found) and unsupported methods
    (405 method_not_allowed with an Allow header), uses the JSON shape
    {"error":{"code":"snake_case","message":"What went wrong and what to do next."}}.
    Request bodies are JSON (Content-Type application/json, optional charset=utf-8);
    decoding errors name the offending field.

    Authentication: sign in with POST /auth/login and send Authorization: Bearer <session_token>.
    Sessions slide (1 hour idle, 12 hours maximum); authenticated responses carry
    X-ShakerProxy-Session-Expires-At. Scoped API tokens (POST /auth/tokens) use the same header.

    Password confirmation: fields named "password" confirm the administrator. After any
    successful password check (including sign-in) the same session may omit "password" for
    10 minutes on routes marked "recent confirmation"; they answer 401 reauthentication_required
    when a password is needed. Routes marked "password every time" (network commit/confirm,
    deletion execution and retention, API token creation, enabling HTTPS decryption for every
    client, enabling decrypted HTTP body retention) answer 401 password_required without one.
servers:
  - url: /api/v1
paths:
...
  /traffic-policy/preview:
    post:
      operationId: previewTrafficPolicy
      description: Validates and renders a proposed policy against the confirmed client interface without mutating host state.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/TrafficPolicy"}}}}
      responses:
        "200": {description: Policy digest, fixed firewall projection, changed objects, and limitations, content: {application/json: {schema: {$ref: "#/components/schemas/TrafficPolicyPreview"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /traffic-policy/rollback:
    post:
      operationId: rollbackTrafficPolicy
      description: Freshly reauthenticated optimistic rollback to the one retained previous policy. Rollback creates a new revision and preserves audit history.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/RollbackTrafficPolicyRequest"}}}}
      responses:
        "200": {description: Newly revisioned rollback result, content: {application/json: {schema: {$ref: "#/components/schemas/TrafficPolicyDocument"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, description: Browser sessions authenticate every route. Expiring scoped API tokens authenticate only explicitly scoped read/automation routes and cannot manage credentials, forwarders, networking, deletion, exports, rules, or trust material.}
  responses:
    Error:
      description: Structured error response
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: {type: string, pattern: "^[a-z0-9_]+$"}
                  message: {type: string, description: What went wrong and what to do next.}
...
  schemas:
...
    EncryptedDNSPolicy:
      type: object
      additionalProperties: false
      required: [mode, block_dot, block_doq, block_known_doh, block_known_doh3, redirect_plain_dns, local_listen_port, upstream_servers]
      properties:
        mode: {type: string, enum: [OBSERVE, BLOCK_KNOWN, ENFORCE_LOCAL]}
        block_dot: {type: boolean}
        block_doq: {type: boolean}
        block_known_doh: {type: boolean}
        block_known_doh3: {type: boolean}
        redirect_plain_dns: {type: boolean, description: Required when mode is ENFORCE_LOCAL.}
        local_listen_port: {type: integer, minimum: 1024, maximum: 65535, default: 1053}
        upstream_servers: {type: array, maxItems: 16, uniqueItems: true, items: {type: string, minLength: 4, maxLength: 64, description: "Canonical IPv4-or-bracketed-IPv6 IP:53 endpoint; hostnames are rejected."}}
        resolver_exclusions: {type: array, maxItems: 512, uniqueItems: true, items: {type: string, pattern: "^[a-z0-9][a-z0-9-]{1,62}$"}}
    TLSInterceptionPolicy:
      type: object
      additionalProperties: false
      required: [enabled, transparent_port, auto_bypass_pinned, auto_bypass_ttl_seconds, max_dynamic_bypasses]
      properties:
        enabled: {type: boolean}
        transparent_port: {type: integer, minimum: 1024, maximum: 65535, default: 8085}
        exclude_hosts: {type: array, maxItems: 512, uniqueItems: true, items: {type: string, maxLength: 253}}
        exclude_cidrs: {type: array, maxItems: 512, uniqueItems: true, items: {type: string, maxLength: 64}}
        auto_bypass_pinned: {type: boolean}
        auto_bypass_ttl_seconds: {type: integer, minimum: 300, maximum: 2592000}
        max_dynamic_bypasses: {type: integer, minimum: 16, maximum: 10000}
        mobile_clients:
          type: array
          maxItems: 512
          items:
            type: object
            additionalProperties: false
            required: [cidr, platform]
            properties:
              cidr: {type: string, maxLength: 18, description: Canonical IPv4 CIDR such as 192.0.2.40/32.}
              platform: {type: string, enum: [android, android-tv, ios, tvos]}
        selected_device_ids: {type: array, maxItems: 512, uniqueItems: true, items: {type: string, pattern: "^device-[a-f0-9]{32}$"}, description: Empty means every lab client; otherwise only these devices are decrypted.}
        allow_quic: {type: boolean, default: false, description: "Keep UDP/443 open for decrypted clients. By default QUIC is rejected so HTTP/3 apps fall back to decryptable TCP."}
        intercept_http: {type: boolean, default: false, description: Also record plain HTTP (TCP/80) from decrypted clients.}
        intercept_private_destinations: {type: boolean, default: false, description: "Also decrypt connections to RFC 1918, CGNAT, link-local and ULA destinations (off by default: LAN and IoT backends often use self-signed or mutual TLS)."}
...
    TrafficDeviceControl:
      type: object
      additionalProperties: false
      required: [device_id]
      properties:
        device_id: {type: string, pattern: "^device-[a-f0-9]{32}$"}
        hardware_addresses: {type: array, maxItems: 8, items: {type: string, pattern: "^[0-9a-f]{2}(:[0-9a-f]{2}){5}$"}, description: Identity evidence captured from the inventory.}
        addresses: {type: array, maxItems: 16, items: {type: string, maxLength: 64}}
        block_internet: {type: boolean}
        blocked_domains: {type: array, maxItems: 256, items: {type: string, maxLength: 253}}
        bypass_hosts: {type: array, maxItems: 256, items: {type: string, maxLength: 253}}
    TrafficPolicy:
      type: object
      additionalProperties: false
      required: [schema, revision, name, encrypted_dns, tls_interception]
      properties:
        schema: {const: 1}
        revision: {type: integer, minimum: 1}
        name: {type: string, minLength: 1, maxLength: 120}
        encrypted_dns: {$ref: "#/components/schemas/EncryptedDNSPolicy"}
        tls_interception: {$ref: "#/components/schemas/TLSInterceptionPolicy"}
        device_controls: {type: array, maxItems: 512, items: {$ref: "#/components/schemas/TrafficDeviceControl"}, description: "Per-device lab controls, normally managed with /devices/{device}/controls. A PUT without this field keeps the saved controls; an empty list clears them."}
    TrafficPolicyDocument:
      type: object
      additionalProperties: false
      required: [schema, policy, digest, applied_at]
      properties:
        schema: {const: 1}
        policy: {$ref: "#/components/schemas/TrafficPolicy"}
        digest: {type: string, pattern: "^[a-f0-9]{64}$"}
        applied_at: {type: string, format: date-time}
        previous: {$ref: "#/components/schemas/TrafficPolicy"}
    TrafficFirewallPreview:
      type: object
      additionalProperties: false
      required: [filter_rules, nat_rules, needs_input_hook, needs_nat_prerouting_hook, needs_dns_service, needs_mitm_service]
      properties:
        filter_rules: {type: array, maxItems: 4096, items: {type: string, maxLength: 1024}}
        nat_rules: {type: array, maxItems: 4096, items: {type: string, maxLength: 1024}}
        needs_input_hook: {type: boolean}
        needs_nat_prerouting_hook: {type: boolean}
        needs_dns_service: {type: boolean}
        needs_mitm_service: {type: boolean}
        filter_rules_ipv6: {type: array, maxItems: 4096, items: {type: string, maxLength: 1024}}
        nat_rules_ipv6: {type: array, maxItems: 4096, items: {type: string, maxLength: 1024}}
        needs_onboarding_service: {type: boolean}
        unmatched_devices: {type: array, maxItems: 512, items: {type: string}, description: Devices with controls but no known MAC or IP yet.}
    TrafficPolicyPreview:
      type: object
      additionalProperties: false
      required: [policy, digest, firewall, changed_objects, warnings]
      properties:
        policy: {$ref: "#/components/schemas/TrafficPolicy"}
        digest: {type: string, pattern: "^[a-f0-9]{64}$"}
        firewall: {$ref: "#/components/schemas/TrafficFirewallPreview"}
        changed_objects: {type: array, minItems: 2, maxItems: 8, uniqueItems: true, items: {type: string, maxLength: 256}}
        warnings: {type: array, maxItems: 8, items: {type: string, maxLength: 512}}
...
    RollbackTrafficPolicyRequest:
      type: object
      additionalProperties: false
      required: [expected_revision]
      properties:
        expected_revision: {type: integer, minimum: 1}
        password: {type: string, minLength: 1, maxLength: 1024}
...